refactor(RVCV-71): 변수타입은 Long > Integer로 변경한다.

+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/shopadmin/application/usecase/ShopAdminUseCase.java
@@ -19,5 +19,5 @@ public interface ShopAdminUseCase {
 
 	List<ReviewDesign> getGeneralReviewThemeList();
 
-	void updateReviewDesign(Long reviewDesignId, UpdateReviewDesignRequest updateReviewDesignRequest);
+	void updateReviewDesign(Integer reviewDesignId, UpdateReviewDesignRequest updateReviewDesignRequest);
 }
